refactor(web): insert all note graphql queries/mutations into dedicated files
