+++ apps/web/src/features/notes/hooks/use-note-query.ts
@@ -1,22 +1,10 @@
 import { useQuery } from "urql";
 
-import { graphql } from "@/gql";
-
-const NoteQuery = graphql(`
-  query Note($id: String!) {
-    note(id: $id) {
-      id
-      title
-      content
-      updatedAt
-      isLocked
-    }
-  }
-`);
+import { NOTE_QUERY } from "@/lib/gql/queries";
 
 export const useNoteQuery = (id: string) => {
   return useQuery({
-    query: NoteQuery,
+    query: NOTE_QUERY,
     variables: { id },
   });
 };
